+++ .changeset/metal-hairs-join.md
@@ -1,0 +1,5 @@
+---
+'@soundxyz/sdk': minor
+---
+
+export Sound API
