+++ packages/sdk/src/graphql.ts
@@ -1,0 +1,16 @@
+import { z } from 'zod'
+
+export const graphqlRequestBody = z.object({
+  data: z.record(z.unknown()).nullable().optional(),
+  errors: z
+    .array(
+      z.object({
+        message: z.string(),
+        locations: z.array(z.object({ line: z.number(), column: z.number() })).optional(),
+        path: z.array(z.union([z.string(), z.number()])).optional(),
+        extensions: z.record(z.unknown()).optional(),
+      }),
+    )
+    .optional(),
+  extensions: z.record(z.unknown()).optional(),
+})
