+++ packages/sdk/src/index.ts
@@ -1,5 +1,9 @@
 export { SoundClient } from './client'
+export { SoundAPI } from './api'
 export { SoundSubgraph } from './subgraph'
+
+export type { SoundAPIConfig } from './api'
+export type { SoundSubgraphConfig } from './subgraph'
 
 export * as Errors from './errors'
 export * from './types'
